+++ python/egglog/__init__.py
@@ -5,7 +5,9 @@ Package for creating e-graphs in Python.
 from . import config, ipython_magic  # noqa: F401
 from .bindings import EggSmolError  # noqa: F401
 from .builtins import *  # noqa: UP029
-from .conversion import ConvertError, convert, converter, get_type_args  # noqa: F401
+from .conversion import *
+from .deconstruct import *
 from .egraph import *
+from .runtime import define_expr_method as define_expr_method
 
 del ipython_magic