+++ python/tests/conftest.py
@@ -3,16 +3,23 @@ import copy
 import pytest
 from syrupy.extensions.single_file import SingleFileSnapshotExtension
 
+import egglog.conversion
+import egglog.exp.array_api
+
 
 @pytest.fixture(autouse=True)
 def _reset_conversions():
-    import egglog.conversion
-
     old_conversions = copy.copy(egglog.conversion.CONVERSIONS)
     old_conversion_decls = copy.copy(egglog.conversion._TO_PROCESS_DECLS)
     yield
     egglog.conversion.CONVERSIONS = old_conversions
     egglog.conversion._TO_PROCESS_DECLS = old_conversion_decls
+
+
+@pytest.fixture(autouse=True)
+def _reset_current_egraph():
+    yield
+    egglog.exp.array_api._CURRENT_EGRAPH = None
 
 
 class PythonSnapshotExtension(SingleFileSnapshotExtension):